refactor: extract theme selector and user menu
