+++ web/src/components/ChatboxHeader/index.module.css
@@ -17,70 +17,3 @@
     margin-left: auto;
     margin-right: 2vw;
 }
-
-.themeMenuTitle {
-    background-color: transparent;
-    color: var(--text-primary);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border: none;
-}
-
-.themeMenuTitle:hover {
-    background-color: var(--bg-secondary);
-    border-radius: 5px;
-}
-
-.themeMenuText {
-    margin: 0.3em;
-    text-align: start;
-    font-size: 0.8rem;
-}
-
-.themeMenuList {
-    background-color: var(--bg-primary);
-    border: 1px solid var(--border-color);
-    border-radius: 3px;
-    padding: 0.5rem;
-}
-
-.themeMenuItem {
-    width: 100%;
-    margin: 2px;
-    background-color: transparent;
-    /* must add transparent border or the button will shake on hover */
-    border: 1px solid transparent;
-    color: var(--text-primary);
-    display: flex;
-    align-items: center;
-}
-
-.themeMenuItem:hover {
-    border: 1px solid var(--border-color);
-    border-radius: 3px;
-}
-
-.themeMenuItem.selected {
-    border: 1px solid var(--border-color);
-    border-radius: 3px;
-}
-
-.userInfoMenu {
-    background-color: transparent;
-    border: none;
-}
-
-.userInfoMenuList {
-    background-color: var(--bg-primary);
-    border: 1px solid var(--border-color);
-    border-radius: 3px;
-    position: absolute;
-    right: 2vw;
-    z-index: 1;
-    padding: 0.5rem;
-}
-
-.userInfoMenuUsernameHr {
-    border: 1px solid var(--border-color);
-}
