+++ web/src/components/ThemeSelector/index.module.css
@@ -1,0 +1,47 @@
+.themeMenuTitle {
+    background-color: transparent;
+    color: var(--text-primary);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border: none;
+}
+
+.themeMenuTitle:hover {
+    background-color: var(--bg-secondary);
+    border-radius: 5px;
+}
+
+.themeMenuText {
+    margin: 0.3em;
+    text-align: start;
+    font-size: 0.8rem;
+}
+
+.themeMenuList {
+    background-color: var(--bg-primary);
+    border: 1px solid var(--border-color);
+    border-radius: 3px;
+    padding: 0.5rem;
+}
+
+.themeMenuItem {
+    width: 100%;
+    margin: 2px;
+    background-color: transparent;
+    /* must add transparent border or the button will shake on hover */
+    border: 1px solid transparent;
+    color: var(--text-primary);
+    display: flex;
+    align-items: center;
+}
+
+.themeMenuItem:hover {
+    border: 1px solid var(--border-color);
+    border-radius: 3px;
+}
+
+.themeMenuItem.selected {
+    border: 1px solid var(--border-color);
+    border-radius: 3px;
+}
